+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/core/refactoring/AddActionInfo.java
@@ -3,7 +3,14 @@ package org.objectstyle.wolips.wodclipse.core.refactoring;
 import org.eclipse.jdt.core.IType;
 import org.eclipse.jdt.core.JavaModelException;
 import org.objectstyle.wolips.bindings.utils.BindingReflectionUtils;
+import org.objectstyle.wolips.variables.ParsleyProject;
 
+/**
+ * Holds the input values for the "Add Action" dialog: name and return type.
+ *
+ * <p>The default return type is determined by the project type: NG projects
+ * default to {@code NGActionResults}, WO projects to {@code WOActionResults}.
+ */
 public class AddActionInfo {
   private IType _componentType;
 
@@ -14,7 +21,24 @@ public class AddActionInfo {
   public AddActionInfo(IType componentType) {
     _componentType = componentType;
     _name = "newAction";
-    _typeName = "WOActionResults";
+    _typeName = resolveDefaultActionResultsType(componentType);
+  }
+
+  /**
+   * Returns "NGActionResults" for ng-objects projects, "WOActionResults" otherwise.
+   * Falls back to "WOActionResults" if the project type cannot be determined.
+   */
+  private static String resolveDefaultActionResultsType(IType componentType) {
+    try {
+      ParsleyProject pp = (ParsleyProject) componentType.getJavaProject().getProject().getAdapter(ParsleyProject.class);
+      if (pp != null && pp.isNGProject()) {
+        return "NGActionResults";
+      }
+    }
+    catch (Exception e) {
+      // Fall through to WO default
+    }
+    return "WOActionResults";
   }
 
   public String getJavaTypeName() throws JavaModelException {